refactored Confluence class into its own file
